:truck: rename css files for consistency
