+++ app/components/machineslider/styles.css
