+++ app/components/parts/genre/index.js
@@ -1,5 +1,5 @@
 import html from 'choo/html';
-import styles from './style.css';
+import styles from './styles.css';
 
 // Import genre icons
 import genreChildrenslife from './icons/genre-childrenslife.svg';
