+++ app/routes/main.js
@@ -3,7 +3,7 @@ import header from '../components/header';
 import machineslider from '../components/machineslider';
 import machineForm from '../components/machineform';
 import modal from '../components/settings-modal';
-import bodyStyles from './body.css';
+import bodyStyles from './main.css';
 import machine from '../components/machine';
 import results from '../components/results';
